Finished servo code, added minor changes to teleop code to make it handle better.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TuxMotor.java
@@ -46,7 +46,7 @@ public class TuxMotor {
         return motor.getCurrentPosition();
     }
 
-    public void setPower(float power) {
+    public void setPower(double power) {
         if(motor.getMode() != DcMotor.RunMode.RUN_WITHOUT_ENCODER) {
             motor.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
         }
